+++ R/is_beast2_installed.R
@@ -4,7 +4,7 @@
 #'   The name of the BEAST2 binary file will be at
 #'   \code{[folder_name]/beast/bin/beast}
 #'   The name of the BEAST2 jar file will be at
-#'   \code{[folder_name]/beast/lib/beast.jar}
+#'   \code{[folder_name]/beast/lib/launcher.jar}
 #' @return TRUE if BEAST2 is installed
 #' @author Richèl J.C. Bilderbeek
 #' @examples
@@ -17,9 +17,9 @@ is_beast2_installed <- function(
   os = rappdirs::app_dir()$os
 ) {
   check_os(os = os) # nolint internal function
-  jar_file_path <- file.path(folder_name, "BEAST", "lib", "beast.jar")
+  jar_file_path <- file.path(folder_name, "BEAST", "lib", "launcher.jar")
   if (os != "win") {
-    jar_file_path <- file.path(folder_name, "beast", "lib", "beast.jar")
+    jar_file_path <- file.path(folder_name, "beast", "lib", "launcher.jar")
   }
   bin_file_path <- file.path(folder_name, "BEAST", "BEAST.exe")
   if (os != "win") {