+++ R/is_jar_path.R
@@ -7,12 +7,12 @@
 #'   library(testthat)
 #'
 #'   expect_false(is_jar_path("beast"))
-#'   expect_true(is_jar_path("beast.jar"))
+#'   expect_true(is_jar_path("launcher.jar"))
 #'   expect_false(is_jar_path(get_default_beast2_bin_path()))
 #'   expect_true(is_jar_path(get_default_beast2_jar_path()))
 #' @noRd
 is_jar_path <- function(path) {
   !beautier::is_one_na(stringr::str_match(
-    path, "(.*/)?beast\\.jar$")[1][1]
+    path, "(.*/)?launcher\\.jar$")[1][1]
   )
 }