+++ R/uninstall_beast2.R
@@ -4,7 +4,7 @@
 #'   The name of the BEAST2 binary file will be at
 #'   \code{[folder_name]/beast/bin/beast}
 #'   The name of the BEAST2 jar file will be at
-#'   \code{[folder_name]/beast/lib/beast.jar}
+#'   \code{[folder_name]/beast/lib/launcher.jar}
 #' @author Richèl J.C. Bilderbeek
 #' @examples
 #'   library(testthat)
@@ -30,9 +30,9 @@ uninstall_beast2 <- function(
     stop("Cannot uninstall absent BEAST2 at", folder_name)
   }
   check_os(os = os) # nolint internal function
-  jar_file_path <- file.path(folder_name, "BEAST", "lib", "beast.jar")
+  jar_file_path <- file.path(folder_name, "BEAST", "lib", "launcher.jar")
   if (os != "win") {
-    jar_file_path <- file.path(folder_name, "beast", "lib", "beast.jar")
+    jar_file_path <- file.path(folder_name, "beast", "lib", "launcher.jar")
   }
   bin_file_path <- file.path(folder_name, "BEAST", "BEAST.exe")
   if (os != "win") {