+++ tests/testthat/test-create_beast2_run_cmd.R
@@ -46,7 +46,6 @@ test_that("use, jar", {
   expected <- c(
     get_default_java_path(),
     "-jar",
-    "-Dbeast.load.jars",
     beast2_jar_path,
     "-statefile",
     paste0("\"", output_state_filename, "\""),