+++ tests/testthat/test-get_default_beast2_jar_path.R
@@ -5,7 +5,7 @@ test_that("must have a correct location", {
   if (rappdirs::app_dir()$os == "unix") {
     testthat::expect_true(
       grepl(
-        "/home/[A-Za-z0-9_]*/.local/share/beast/lib/beast.jar", # nolint use an absolute path here
+        "/home/[A-Za-z0-9_]*/.local/share/beast/lib/launcher.jar", # nolint use an absolute path here
         get_default_beast2_jar_path()
       )
     )