+++ tests/testthat/test-install_beast2.R
@@ -5,7 +5,7 @@ test_that("install at non-standard location", {
   if (!is_on_ci()) return()
 
   folder_name <- tempdir()
-  beast_jar_path <- file.path(folder_name, "beast", "lib", "beast.jar")
+  beast_jar_path <- file.path(folder_name, "beast", "lib", "launcher.jar")
   testit::assert(!file.exists(beast_jar_path))
   testthat::expect_output(
     install_beast2(folder_name, verbose = TRUE),